refactor(storage): extract branch primitives to branch-index.ts (#2106 R10)
